+++ src/main/java/janggi/domain/position/Column.java
@@ -1,4 +1,4 @@
-package janggi.domain;
+package janggi.domain.position;
 
 public record Column(
         int column
