+++ src/main/java/janggi/domain/position/Position.java
@@ -1,8 +1,8 @@
-package janggi.domain;
+package janggi.domain.position;
 
 public record Position(
         Row row,
         Column column
 ) {
-    
+
 }
